requirements.txt
certifi==2018.4.16
dj-database-url==0.5.0
Django==2.0.4
django-bower==5.2.0
django-heroku==0.3.1
django-scheduler==0.8.8
gunicorn==19.8.1
google-api-python-client==1.7.8
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
icalendar==4.0.1
psycopg2==2.7.4
python-dateutil==2.7.2
pytz==2018.4
six==1.11.0
whitenoise==3.3.1
wincertstore==0.2
